backend/requirements.txt
fastapi>=0.115.4
uvicorn[standard]>=0.30.0
psycopg2-binary>=2.9.9
prometheus-fastapi-instrumentator>=7.0.0
wheel>=0.46.2
jaraco.context>=6.1.0